requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
Pillow
numpy
pyinstaller
